src/engined/pyproject.toml
[project]
name = "sigmavault-engined"
version = "0.1.0"
description = "SigmaVault NAS OS - AI/ML RPC Engine with 40-Agent Swarm"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.11"
authors = [
    { name = "SigmaVault Team" }
]
keywords = [
    "nas",
    "ai",
    "compression",
    "encryption",
    "quantum-resistant",
    "grpc",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    
    # gRPC
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "protobuf>=5.29.0",
    
    # Data Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    
    # AI/ML Core
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
    "torch>=2.5.0",
    "transformers>=4.47.0",
    
    # Compression
    "zstandard>=0.23.0",
    "lz4>=4.3.0",
    "brotli>=1.1.0",
    
    # Cryptography (Quantum-Resistant)
    "cryptography>=44.0.0",
    "pynacl>=1.5.0",
    "liboqs-python>=0.11.0",
    
    # Storage & ZFS
    "pyudev>=0.24.0",
    
    # Async & Concurrency
    "asyncio-pool>=0.7.0",
    "anyio>=4.7.0",
    
    # Observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-grpc>=0.50b0",
    
    # Configuration
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    
    # Redis for caching/pubsub
    "redis>=5.2.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "pre-commit>=4.0.0",
    "hypothesis>=6.120.0",
    "grpcio-testing>=1.68.0",
]

docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
]

[project.scripts]
sigmavault-engined = "engined.main:main"
sigmavault-agent = "engined.cli:agent_cli"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engined"]

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "transformers.*",
    "grpc.*",
    "liboqs.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C4",     # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["engined"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

[tool.coverage.run]
branch = true
source = ["engined"]
omit = ["tests/*", "**/migrations/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
